Revert "[directory cache] Fixed dram directory cache problem"
This reverts commit 1867a6a65cd12de76730daf1dfd5d54486ba36c7.

+++ common/system/vm_manager.cc
@@ -34,8 +34,8 @@ VMManager::VMManager()
    UInt32 stack_size_per_core;
    try
    {
-      m_start_stack_segment = boost::lexical_cast <IntPtr> (Sim()->getCfg()->get("stack/stack_base"));
-      stack_size_per_core = boost::lexical_cast <UInt32> (Sim()->getCfg()->get("stack/stack_size_per_core"));
+      m_start_stack_segment = (IntPtr) boost::lexical_cast <unsigned long int> (Sim()->getCfg()->get("stack/stack_base"));
+      stack_size_per_core = (UInt32) boost::lexical_cast <unsigned long int> (Sim()->getCfg()->get("stack/stack_size_per_core"));
    }
    catch (boost::bad_lexical_cast &)
    {
